tec: Bump Ruby version to 2.6.3
Travis was failing because it was trying to write `Gemfile.lock`
(through the Jekyll Docker image). The reason of that is the Jekyll
image runs on Ruby 2.6.3 and so it wants to update the lock file. By
upgrading the supported Ruby version, Travis doesn't need to write this
file anymore.